refactor: userinfo
