+++ apps/lostark-hands-next/app/user-info/[name]/@component/card-set/skeleton.tsx
@@ -1,0 +1,18 @@
+import { LabelLayoutSkeleton } from '@/client-component/label-layout';
+import Skeleton from '@/client-component/skeleton';
+
+const CardSetSkeleton = () => (
+	<LabelLayoutSkeleton as="section">
+		<div className="grid cursor-pointer grid-cols-6 gap-[8px]">
+			{Array.from({ length: 6 }).map((_, idx) => (
+				<Skeleton
+					key={idx}
+					className="w-full pb-[146%]"
+					type="LIGHT"
+				/>
+			))}
+		</div>
+	</LabelLayoutSkeleton>
+);
+
+export default CardSetSkeleton;
