+++ apps/lostark-hands-next/app/user-info/[name]/@component/combat-skill/types.ts
@@ -1,0 +1,3 @@
+import type { skillSelector } from '@/service/armories/selector';
+
+export type TData = ReturnType<typeof skillSelector>;
