+++ apps/lostark-hands-next/app/user-info/[name]/@component/equipment/skeleton.tsx
@@ -1,0 +1,44 @@
+import { LabelLayoutSkeleton } from '@/client-component/label-layout';
+import { ThumbnailCardSkeleton } from '@/client-component/thumbnail-card';
+
+const EquipmentSkeleton = () => (
+	<>
+		<LabelLayoutSkeleton
+			className="!mt-0"
+			as="section"
+		>
+			<div className="grid grid-cols-1 gap-[8px] lg:grid-cols-2">
+				{Array.from({ length: 6 }).map((_, idx) => (
+					<div
+						className="flex space-x-[6px]"
+						key={idx}
+					>
+						<ThumbnailCardSkeleton className="h-[40px] w-[40px]" />
+					</div>
+				))}
+			</div>
+		</LabelLayoutSkeleton>
+		<LabelLayoutSkeleton as="section">
+			<div className="grid grid-cols-1 gap-[8px] lg:grid-cols-2">
+				<div className="flex flex-col space-y-[8px]">
+					{Array.from({ length: 5 }).map((_, idx) => (
+						<ThumbnailCardSkeleton
+							className="h-[40px] w-[40px]"
+							key={idx}
+						/>
+					))}
+				</div>
+				<div className="flex flex-col space-y-[8px]">
+					{Array.from({ length: 2 }).map((_, idx) => (
+						<ThumbnailCardSkeleton
+							className="h-[40px] w-[40px]"
+							key={idx}
+						/>
+					))}
+				</div>
+			</div>
+		</LabelLayoutSkeleton>
+	</>
+);
+
+export default EquipmentSkeleton;
