+++ apps/lostark-hands-next/app/user-info/[name]/@component/equipment/types.ts
@@ -1,0 +1,9 @@
+import type {
+	ISelectedArmoryEquipment,
+	TParsedArmory
+} from '@/service/armories/types';
+
+export interface IEquipCardProps
+	extends TParsedArmory<ISelectedArmoryEquipment> {
+	onClick: () => void;
+}
