+++ apps/lostark-hands-next/app/user-info/[name]/@component/profile/constants.ts
@@ -1,0 +1,1 @@
+export const TEXT_STYLE = 'text-[14px] md:text-[16px]';
