+++ apps/lostark-hands-next/app/user-info/[name]/@component/stats/index.tsx
@@ -7,7 +7,6 @@ import type { IStat } from '@/service/armories/types';
 import Label from '@/client-component/label';
 import { LabelLayout } from '@/client-component/label-layout';
 import DescListModal from '@/client-component/modal/desc-list-modal';
-import Skeleton from '@/client-component/skeleton';
 
 interface IStatsProps {
 	data: {
@@ -15,7 +14,7 @@ interface IStatsProps {
 	};
 }
 
-export const Stats = ({ data: { stats: initStats } }: IStatsProps) => {
+const Stats = ({ data: { stats: initStats } }: IStatsProps) => {
 	const { onOpenModal } = useModal();
 
 	const stats = initStats.slice(0, 6);
@@ -94,49 +93,4 @@ export const Stats = ({ data: { stats: initStats } }: IStatsProps) => {
 	);
 };
 
-export const StatsSkeleton = () => (
-	<Skeleton className="!mt-0 p-[8px]">
-		<div className="grid grid-cols-2 gap-[8px]">
-			<div className="space-y-[4px] text-center">
-				<Skeleton
-					type="LIGHT"
-					className="h-[29px] w-full"
-				/>
-				<Skeleton
-					type="LIGHT"
-					className="mx-auto h-[21px]"
-					randomWidth={{ max: 60, min: 40 }}
-				/>
-			</div>
-			<div className="space-y-[4px] text-center">
-				<Skeleton
-					type="LIGHT"
-					className="h-[29px] w-full"
-				/>
-				<Skeleton
-					type="LIGHT"
-					className="mx-auto h-[21px]"
-					randomWidth={{ max: 60, min: 40 }}
-				/>
-			</div>
-		</div>
-		<div className="mt-[12px] grid grid-cols-2 gap-[6px]">
-			{Array.from({ length: 6 }).map((_, idx) => (
-				<div
-					className="flex items-center space-x-[4px]"
-					key={idx}
-				>
-					<Skeleton
-						type="LIGHT"
-						className="h-[29px] w-[36px]"
-					/>
-					<Skeleton
-						type="LIGHT"
-						className="h-[29px]"
-						randomWidth={{ max: 36, min: 16 }}
-					/>
-				</div>
-			))}
-		</div>
-	</Skeleton>
-);
+export default Stats;
